moved everything regarding the Tracking logic to own folder /track
